Added some output filtering; Midi LUT saved to Flash directly
Changed sample rate to 44kHz

+++ DisPhazePoly/src/MidiHandler.h
@@ -26,7 +26,7 @@ public:
 		uint8_t origNote;		// MIDI note value
 		float noteValue;		// MIDI note value adjusted with pitch bend
 		uint8_t envelope;		// Stage in envelope
-		int32_t envTime;		// envelope sample counter
+		float envTime;		// envelope sample counter
 		float samplePos1;		// Current position within cycle
 		float samplePos2;
 		float releaseLevel;		// Store level when hitting release phase
